Add GenCRepr and GenCMsg libraries
These are generic C codegen libraries for data interop with C and credit-based messaging.

GenCRepr provides generic byte-aligned versions of pack/unpack, and can be used to generate a C library at elaboration time for packing and unpacking of the same data to and from an equivalent struct/tagged union representation.

GenCMsg uses GenCRepr to generically implement messaging protocols with credit-based control to avoid ahead-of-line blocking. Given a message interface, generics are used to synthesize an implementation Bluespec module and generate a corresponding C library.